% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.assam.R
\name{fitted.assam}
\alias{fitted.assam}
\title{Extract fitted values from an approximate and scalable SAM object}
\usage{
\method{fitted}{assam}(object, type = "all", ...)
}
\arguments{
\item{object}{An object of class \code{assam}.}

\item{type}{The type of fitted value to obtain. Currently one of three options are possible: \code{all}, which for each species returns the fitted values for all archetypes; \code{max}, which for each species returns the fitted values corresponding to the most likely archetype the species belongs to (as based on the posterior probabilities); \code{mean}, which for each species returns a weighted mean of the fitted values from each archetype, where the weights are given by the posterior probabilities.}

\item{...}{Not used.}
}
\value{
A matrix of fitted values.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Extracts fitted mean values from a \code{assam} object.
}
\details{
To clarify, the returned fitted values are on the response scale i.e., a matrix of the estimated means \eqn{\hat{\mu}_{ij}} after model fitting.
}
\examples{
\dontrun{
#' Please see the help file for assam for example.
}

}
\author{
Francis K.C. Hui \href{mailto:fhui28@gmail.com}{fhui28@gmail.com}
}
