% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lseq.R
\name{lseq}
\alias{lseq}
\title{Produce a sequence of tuning parameters on the log (base 10) scale.}
\usage{
lseq(from, to, length, decreasing = FALSE)
}
\arguments{
\item{from}{Minimum tuning parameter in sequence.}

\item{to}{Maximum tuning parameter in sequence.}

\item{length}{Length of tuning parameter sequence.}

\item{decreasing}{Should the sequence of tuning parameters be increasing or decreasing? Defaults to \code{FALSE} i.e., increasing.}
}
\value{
A sequence of tuning parameters.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\author{
Francis K.C. Hui \href{mailto:fhui28@gmail.com}{fhui28@gmail.com}
}
