% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.assam.R
\name{plot.assam}
\alias{plot.assam}
\title{Basic residual diagnostic plots for an approximate and scalable SAM object}
\usage{
\method{plot}{assam}(
  x,
  y,
  which_plot = 1:5,
  type = "dunnsmyth",
  transform_fitted_values = FALSE,
  titles = c("Residuals vs. fitted values", "Normal probability plot",
    "Residuals vs. unit index", "Residuals vs. species index", "Scale-Location plot"),
  species_colors = NULL,
  smooth = TRUE,
  envelope = TRUE,
  envelope_col = c("grey", "lightgrey"),
  envelope_rep = 100,
  which_species = NULL,
  seed = NULL,
  ...
)
}
\arguments{
\item{x}{An object of class \code{assam}.}

\item{y}{The multivariate abundance data used as part of the \code{assam} object. This needs to be supplied since currently \code{assam} objects do not save the responses to save memory.}

\item{which_plot}{If a subset of the plots is desired, then a vector containing subset of the integers 1, 2, 3, 4, 5.}

\item{type}{The type of residuals to be used in constructing the plots. Currently the options available are: "PIT" and "dunnsmyth".}

\item{transform_fitted_values}{For plots 1 and 5, should the fitted values be transformed using the link function specified in the asSAM i.e.. what is available in \code{x$family$linkfun}. This can be useful visually so that observations are not too bunched up on the x-axis of plots 1 and 5.}

\item{titles}{Titles to appear above each plot.}

\item{species_colors}{Either a scalar is supplied or a vector with length of number of species in the spatio-temporal multivariate abundance data. If the former than all species use this color in the plots. If the latter then the vector specified the colors to use for each species. Defaults to \code{NULL}, which results in each species having a unique color based on the \code{\link[grDevices:palettes]{grDevices::rainbow()}} palette.}

\item{smooth}{Should a smoother be added to each plot?}

\item{envelope}{Should approximate simulation envelopes be constructed for the normal probability plot? Default to \code{TRUE}. Note if \code{envelope = FALSE} then \code{envelope_col} and \code{envelope_K} are ignored.}

\item{envelope_col}{A vector of length 2, specifying the colors to use for the lines and shade respectively for the approximate simulation envelopes.}

\item{envelope_rep}{The number of simulations to use in order to build the approximate simulation envelope.}

\item{which_species}{A vector indexing the species to plot, if the residual plots should be constructed for only a subset of species. Defaults to \code{NULL}, in which case all species are plotted. This may be useful if the number of species is quite large.}

\item{seed}{This can be used set the seed when constructing the PIT and Dunn-Smyth residuals, which for discrete responses involve some degree of jittering.}

\item{...}{Additional graphical arguments.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Five potential plots are currently available for some basic residual diagnostics for a fitted \code{assam} model: 1) a plot of residuals against (potentially transformed) fitted values; 2) a normal probability plot of residuals with simulated point-wise 95\\% confidence interval envelope; 3) plot of residuals against observational unit index; 4) a plot of residuals again column index; 5) scale-location plot using (potentially transformed) fitted values.
}
\details{
This function is heavily adapted from \code{\link[CBFM:plot.CBFM]{CBFM::plot.CBFM()}} and \code{\link[gllvm:plot.gllvm]{gllvm::plot.gllvm()}}. As basic visual diagnostics, these plots should behave as follows:
\enumerate{
\item the plot of residuals versus fitted values should not exhibit any noticeable pattern e.g., no (inverse) fan-shape or a trend;
\item the normal probability plot should have the residuals lying approximately on a straight line and almost all residuals lying within the approximate simulation envelopes;
\item a plot of residuals against observational unit index should not exhibit any noticeable pattern. The plot can also be used to look for potential outlying observational units;
\item a plot of residuals against species index should not exhibit any noticeable pattern. The plot can also be used to look for potential outlying species;
\item the scale-location plot should not exhibit any trend.
}
}
\examples{
\dontrun{
#' Please see the help file for assam for example.
}

}
\author{
Francis K.C. Hui \href{mailto:fhui28@gmail.com}{fhui28@gmail.com}
}
