% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.assam.R
\name{print.assam}
\alias{print.assam}
\title{Print a (approximate and scalable) SAM object}
\usage{
\method{print}{assam}(x, ...)
}
\arguments{
\item{x}{An object of class \code{assam}.}

\item{...}{Not used.}
}
\description{
`r lifecycle::badge("stable")

The default print method for a \code{assam} object.
}
\details{
Print out details such as the function call, assumed family/response distribution, number of sites and species, and archetypal response-environment relationship fitted.
}
\examples{
\dontrun{
#' Please see the help file for assam for example.
}

}
\author{
Francis K.C. Hui \href{mailto:fhui28@gmail.com}{fhui28@gmail.com}
}
