% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.passam.R
\name{print.passam}
\alias{print.passam}
\title{Print a approximate and scalable SAM (pasSAM) object}
\usage{
\method{print}{passam}(x, ...)
}
\arguments{
\item{x}{An object of class \code{passam}.}

\item{...}{Not used.}
}
\description{
`r lifecycle::badge("stable")

The default print method for a \code{passam} object.
}
\details{
Print out details such as the function call, assumed family/response distribution, number of sites and species, and archetypal response-environment relationship fitted.
}
\examples{
\dontrun{
#' Please see the help file for assam and passam for example.
}

}
\author{
Francis K.C. Hui \href{mailto:fhui28@gmail.com}{fhui28@gmail.com}
}
