% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.assam.R
\name{residuals.assam}
\alias{residuals.assam}
\title{Extract residuals from an approximate and scalable SAM object}
\usage{
\method{residuals}{assam}(object, y, type = "dunnsmyth", seed = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{assam}.}

\item{y}{The multivariate abundance data used as part of the \code{assam} object. This needs to be supplied since currently \code{assam} objects do not save the responses to save memory.}

\item{type}{The type of residuals which should be returned. The options currently available are "PIT" and "dunnsmyth".}

\item{seed}{This can be used set the seed when constructing the PIT and Dunn-Smyth residuals, which for discrete responses involve some degree of jittering.}

\item{...}{Not used.}
}
\value{
A matrix of residuals.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Calculate either probability integral transform (PIT) residuals and Dunn-Smyth residuals residuals from a fitted \code{assam} object.
}
\details{
Of the two options available \code{type = "PIT"} returns the probability integral transform residuals that are also used in \code{\link[DHARMa:simulateResiduals]{DHARMa::simulateResiduals()}} and \code{\link[mpcmp:rPIT]{mpcmp::rPIT()}}, among other packages. If the (estimated) model is correct, then these residuals should behave as random variables from a standard uniform distribution \href{https://doi.org/10.2307/1390802}{Dunn and Smyth, 1996}. Note there is a level of jittering used in producing the PIT residuals. On the other hand, \code{type = "dunnsmyth"} returns the Dunn-Smyth residuals that are also used in \code{\link[boral:ds.residuals]{boral::ds.residuals()}} and\code{\link[gllvm:residuals.gllvm]{gllvm::residuals.gllvm()}}, among other packages. If the (estimated) model is correct, then these residuals should behave as random variables from a standard normal distribution (Dunn and Smyth, 1996). Note there is a level of jittering used in producing the Dunn-Smyth residuals.

Both PIT and Dunn-Smyth residuals were adapted by \href{https://doi.org/10.1007/s13253-013-0146-x}{Dunstan et al., (2013)} to species archetype models, and we follow their developments in building this function.
}
\examples{
\dontrun{
#' Please see the help file for assam for example.
}

}
\author{
Francis K.C. Hui \href{mailto:fhui28@gmail.com}{fhui28@gmail.com}
}
