% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.assam.R
\name{simulate.assam}
\alias{simulate.assam}
\alias{simulate}
\title{Simulate data from approximate and scalable SAM object}
\usage{
\method{simulate}{assam}(object, data, nsim = 1, seed = NULL, ...)

simulate(object, ...)
}
\arguments{
\item{object}{An object of class \code{assam}.}

\item{data}{The data frame used as part of the \code{assam} object. This needs to be supplied since currently \code{assam} objects do not save the data to save memory. Having said that, a user can in principle supply a new data frame to \code{data}, all the other arguments and parameter estimates from the \code{assam} object are compatible.}

\item{nsim}{A positive integer specifying the number of simulated datasets. Defaults to 1.}

\item{seed}{An integer to set seed number. Defaults to a random seed number.}

\item{...}{Not used.}
}
\value{
A matrix with two rows and \code{nsim} columns, where the first row contains the simulated multivariate abundance response matrix, and the second row contains the vector of archetype labels for each species.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Simulate multivariate abundance data based on a fitted \code{assam} object.
}
\details{
Simulates multivariate abundance data from a species archetype model (SAM). For the purposes of the package, the SAM is characterized by the following mean regression model: for observational unit \eqn{i=1,\ldots,N} and species \eqn{j=1,\ldots,M}, conditional on the species belong to archetype \eqn{k},

\deqn{g(\mu_{ij}) = \eta_{ij} = x_i^\top\beta_k,}

where \eqn{g(.)} is a known link function, \eqn{x_i} denotes a vector of predictors for unit \eqn{i} i.e., the \eqn{i}-th row from the created model matrix, \eqn{\beta_k} denotes the corresponding regression coefficients for archetype \eqn{k}. Based on the mean model given above, responses \eqn{y_{ij}} are then simulated from the assumed distribution, using the additional dispersion and power parameters as appropriate.
}
\author{
Francis K.C. Hui \href{mailto:fhui28@gmail.com}{fhui28@gmail.com}
}
