% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.assam.R
\name{summary.assam}
\alias{summary.assam}
\title{Summary for an approximate and scalable SAM object}
\usage{
\method{summary}{assam}(object, digits = 4, ...)
}
\arguments{
\item{object}{An object of class \code{assam}.}

\item{digits}{The number of significant figures to use when printing.}

\item{...}{Not used.}
}
\value{
An object of class "summary.assam" which includes the following components, not necessarily in the order below (and as appropriate):
\describe{
\item{call: }{The matched function call of \code{object}.}
\item{mixture_proportion:}{Estimated vector of mixture proportions corresponding to the probability of belonging to each archetype.}
\item{spp_intercepts:}{Estimated species-specific intercepts.}
\item{betas:}{Estimated matrix of archetypal regression coefficients corresponding to the model matrix created. The number of rows in \code{betas} is equal to the number of archetypes.}
\item{betas_results: }{If the \code{object$uncertainty_quantification = TRUE}, a summary table corresponding to archetypal regression coefficients.}
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Takes a fitted \code{assam} object and produces some basic summary information.
}
\details{
Currently, the function returns estimated mixture proportions, species-specific intercepts, and archetypal regression coefficients. If the \code{object$uncertainty_quantification == TRUE}, then summary tables are also produced for archetypal regression coefficients containing parametric bootstrap confidence intervals and statistical significance based off this. Note the length of the confidence intervals was determined as part of the fitting the asSAM; see \code{object$bootstrap_control$ci_alpha}.
}
\section{Warning}{
The current summary function is pretty basic (apologies!), and in the future we may add some more useful information to the summary output.
}

\author{
Francis K.C. Hui \href{mailto:fhui28@gmail.com}{fhui28@gmail.com}
}
