% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.assam.R
\name{predict.assam}
\alias{predict.assam}
\title{Construct predictions from an approximate and scalable SAM object}
\usage{
\method{predict}{assam}(
  object,
  newdata,
  newoffset = NULL,
  type = c("species_max", "species_mean", "archetype"),
  se_fit = FALSE,
  num_cores = NULL,
  coverage = 0.95,
  ...
)
}
\arguments{
\item{object}{An object of class \code{assam}.}

\item{newdata}{A data frame containing the values of the covariates at which predictions are to be calculated, that is, a model matrix from this and \code{object$formula}.}

\item{newoffset}{A set of offset terms. If supplied, then it must either be a vector (if \code{type = "archetype"}) or a matrix (if \code{type = "species_max"} or \code{"species_mean"}). If the former, the length should be equal to \code{nrow(newdata)}. If the latter, the number of rows should be equal \code{nrow(newdata)} and the number of columns should be equal to \code{length(object$spp_intercepts) i.e., the number of species.}}

\item{type}{The type of prediction required:
If \code{type = "archetype"}, then archetypal predictions are constructed on the scale of the linear predictors i.e., \eqn{x_i^\top\beta_k};
If \code{type = "species_max"}, then species-specific predictions on the scale of the responses i.e., \eqn{\mu_{ij}}, are constructed based on the most likely archetype the species belongs to (as based on the posterior probabilities);
If \code{type = "species_mean"}, then species-specific predictions on the scale of the responses i.e., \eqn{\mu_{ij}}, are constructed based on a weighted mean of the predictions from each archetype, where the weights are given by the posterior probabilities.

Note for both types of species-specific predictions, \strong{currently uncertainty of the posterior probabilities are not taken into account...this is on the to-do list!}}

\item{se_fit}{When this is set to \code{TRUE} (not default), then uncertainty intervals are returned for the point predictions.}

\item{num_cores}{To speed up calculation of the uncertainty intervals, parallelization can be performed, in which case this argument can be used to supply the number of cores to use in the parallelization. Defaults to \code{detectCores()-2}.}

\item{coverage}{The coverage probability of the uncertainty intervals for prediction. Defaults to 0.95, which corresponds to 95\% uncertainty intervals.}

\item{...}{Not used.}
}
\value{
If \code{se_fit = FALSE}, then a matrix of point predictions. If \code{se_fit = TRUE}, then a list with the following components is returned:
\item{point_prediction:}{A matrix of predicted values.}
\item{lower:}{A matrix of the lower limits for the uncertainty intervals.}
\item{upper:}{A matrix of the upper limits for the uncertainty intervals.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Takes a fitted \code{assam} object and produces predictions given (potentially) a new set of observational units with their corresponding covariates. Predictions can be accompanied by uncertainty intervals.
}
\details{
Uncertainty intervals produced by \code{predict.assam} are based on the bootstrapped parameter estimates, if available from the \code{object} itself. That is, predictions are constructed using the bootstrapped estimates and then quantiles as appropriate as used to construct the uncertainty intervals.

Note archetypal predictions are constructed on the scale of the linear predictions, while species-specific predictions are constructed on the scale of the responses. The latter is analogous to what is available from \code{\link[=fitted.assam]{fitted.assam()}}.
}
\examples{
\dontrun{
##----------------------
# Generate some multivariate abundance data from a SAM
##----------------------
library(tidyverse)
library(mvtnorm)
library(GGally)

set.seed(092024)

num_X <- 10
num_units <- 1000
num_spp <- 80
num_archetype <- 5
H <- outer(1:num_X, 1:num_X, "-")
H <- 0.5^abs(H)
covariate_dat <- rmvnorm(num_units, sigma = H) \%>\% 
    as.data.frame \%>\% 
    rename_with(., .fn = function(x) paste0("covariate", x))
rm(H)

true_betas <- runif(num_archetype * num_X, -1, 1) \%>\% matrix(nrow = num_archetype)
true_intercepts <- runif(num_spp, -3, 0)  
true_dispparam <- 1/runif(num_spp, 0, 5) 
true_powerparam <- runif(num_spp, 1.4, 1.8)
true_mixprop <- c(0.2, 0.25, 0.3, 0.1, 0.15)
 
 simdat <- create_samlife(family = nb2(), 
 formula = paste("~ ", paste0(colnames(covariate_dat), collapse = "+")) \%>\% as.formula, 
 data = covariate_dat, 
 betas = true_betas, 
 spp_intercept = true_intercepts, 
 spp_dispparam = true_dispparam, 
 spp_powerparam = true_powerparam, 
 mixture_proportion = true_mixprop)
 
 
 ##----------------------
 # Fit asSAM and assess results 
 #' **Most users should start here**
 ##----------------------
 samfit <- assam(y = simdat$y,
 formula = paste("~ ", paste0(colnames(covariate_dat), collapse = "+")) \%>\% as.formula,
 data = covariate_dat,
 family = nb2(),
 num_archetypes = num_archetype,
 num_cores = 8)
 
 #' Archetype-level predictions
 predict(samfit, newdata = covariate_dat, type = "archetype", se_fit = TRUE) 
 
 #' Species-level predictions
 predict(samfit,  newdata = covariate_dat, type = "species_max", num_cores = 8, se_fit = TRUE) 
 
}


}
\author{
Francis K.C. Hui \href{mailto:fhui28@gmail.com}{fhui28@gmail.com}
}
